+++ projects/ng-rule-editor/src/lib/select-scoring-items/select-scoring-items.component.ts
@@ -29,6 +29,9 @@ export class SelectScoringItemsComponent implements OnInit {
   itemList = [];
   selectedItemsSet = new Set<string>();
 
+  matToolTip = "";
+  validationErrorMessage = "";
+
   options: ITreeOptions = {
     displayField: 'text',
     childrenField: 'item',
@@ -99,6 +102,8 @@ export class SelectScoringItemsComponent implements OnInit {
     this.hasChildren = this.hasChildItems(this.scoringItems);
     // If there are child items, if yes then we want to expand the tree by default.
     this.expandAll = this.hasChildren;
+
+    this.setScoreSelectedStatus(false);
   };
 
   /**
@@ -110,6 +115,23 @@ export class SelectScoringItemsComponent implements OnInit {
     return items.some((item) => {
       return item.hasOwnProperty('item') && item.item.length > 0;
     });
+  }
+
+  /**
+   * Set the status of the Scoring Item Selection, validation message, and tooltip
+   * @param status - true if at least one scoring item is selected
+   */
+  private setScoreSelectedStatus(status): void {
+    this.hasScoreSelected = status;
+
+    if (status) {
+      this.matToolTip = "";
+      this.validationErrorMessage = this.matToolTip;
+    } else {
+      this.matToolTip = 
+        "The 'done' button is disabled because at least one scoring item selection is required.";;
+      this.validationErrorMessage = this.matToolTip;
+    }
   }
 
   /**
@@ -129,7 +151,7 @@ export class SelectScoringItemsComponent implements OnInit {
     
     if (this.scoringItems.length > 0) {
       this.itemTree.treeModel.getVisibleRoots().forEach((item) => toggleItemHierarchy(item, status));
-      this.hasScoreSelected = status;
+      this.setScoreSelectedStatus(status);
     }
   }
 
@@ -194,7 +216,7 @@ export class SelectScoringItemsComponent implements OnInit {
 
           // If there are pre-selected items in the questionnaire, then we want
           // to make sure that the "Done" button is enabled.
-          this.hasScoreSelected = true;
+          this.setScoreSelectedStatus(true);
         }
       });
     }    
@@ -223,7 +245,7 @@ export class SelectScoringItemsComponent implements OnInit {
   onScoringItemCheckboxClick(node): void {
     node.toggleActivated(true);
     const count = this.itemTree.treeModel.getActiveNodes().length;
-    this.hasScoreSelected = (count > 0);
+    this.setScoreSelectedStatus((count > 0));
   }
 
 }