+++ src/app/app.component.ts
@@ -53,7 +53,7 @@ export class AppComponent implements OnInit, OnDestroy {
   expressionUri = this.calculatedExpression;
   userExpressionChoices = null;
   customExpressionUri = false;
-  fhir = null;
+  fhirQuestionnaire = null;
   questionnaire = 'bmisimple';
   file = '';
   error = '';
@@ -84,7 +84,7 @@ export class AppComponent implements OnInit, OnDestroy {
 
     if (this.questionnaire === '' || this.questionnaire === 'upload') {
       this.liveAnnouncer.announce('Additional settings must be entered below to load the rule editor.');
-      this.fhir = null;
+      this.fhirQuestionnaire = null;
       this.file = '';
       this.linkId = '';
       this.rootLevel = true;
@@ -95,11 +95,15 @@ export class AppComponent implements OnInit, OnDestroy {
 
       this.http.get(`./${this.questionnaire}.json`)
         .subscribe(data => {
-          this.fhir = data;
+          this.fhirQuestionnaire = data;
+
+          console.log('app::onChange::fhirQuestionnaire - ' + JSON.stringify(this.fhirQuestionnaire));
+
           this.liveAnnouncer.announce((reload) ? this.formReloadAnnouncement : this.formAppearedAnnouncement);
 
-          if (this.fhir && this.fhir.item instanceof Array) {
-            this.linkIds = this.getQuestionnaireLinkIds(this.fhir.item);
+          if (this.fhirQuestionnaire && this.fhirQuestionnaire.item instanceof Array) {
+            this.linkIds = this.getQuestionnaireLinkIds(this.fhirQuestionnaire.item);
+console.log('app::onChange::fhirQuestionnaire::linkIds - ' + this.linkIds);
 
             this.defaultItemText = this.linkIds.find((item) => {
               return item.linkId === this.linkId;
@@ -165,10 +169,10 @@ export class AppComponent implements OnInit, OnDestroy {
           this.doNotAskToCalculateScore = false;
           this.linkId = '';
           try {
-            this.fhir = JSON.parse(e.target.result);
+            this.fhirQuestionnaire = JSON.parse(e.target.result);
             this.error = '';
-            if (this.fhir && this.fhir.item instanceof Array) {
-              this.linkIds = this.getQuestionnaireLinkIds(this.fhir.item);
+            if (this.fhirQuestionnaire && this.fhirQuestionnaire.item instanceof Array) {
+              this.linkIds = this.getQuestionnaireLinkIds(this.fhirQuestionnaire.item);
 
               this.composeAutocomplete();
             }
@@ -179,12 +183,12 @@ export class AppComponent implements OnInit, OnDestroy {
               this.rootLevel = true;
             }
           } catch (e) {
-            this.fhir = '';
+            this.fhirQuestionnaire = '';
             this.error = `Could not parse file: ${e}`;
             this.liveAnnouncer.announce(this.error);
           }
         } else {
-          this.fhir = '';
+          this.fhirQuestionnaire = '';
           this.error = 'Could not read file';
           this.liveAnnouncer.announce(this.error);
         }
@@ -321,32 +325,34 @@ export class AppComponent implements OnInit, OnDestroy {
    * selected item/question
    */
   openRuleEditorDialog(): void {
-    this.displayRuleEditor = true;
-    this.displayRuleEditorResult = false;
-
-    // The lhc-rule-editor component is not presented before the
-    // 'Open Rule Editor' button is clicked due to the use of *ngIf.
-    // The attributes for the lhc-rule-editor component are not 
-    // getting updated as a result. The below steps are used to 
-    // trigger changes to those attributes. 
-    const tmpUserExpressionChoices = this.userExpressionChoices;
-    const tmpCustomExpressionUri = this.customExpressionUri;
- 
-    this.userExpressionChoices = null;
-    this.customExpressionUri = null;
-
-    this.changeDetectorRef.detectChanges();
-
-    this.userExpressionChoices = tmpUserExpressionChoices;
-    this.customExpressionUri = tmpCustomExpressionUri;
+    if (this.canOpenRuleEditor()) {
+      this.displayRuleEditor = true;
+      this.displayRuleEditorResult = false;
+  
+      // The lhc-rule-editor component is not presented before the
+      // 'Open Rule Editor' button is clicked due to the use of *ngIf.
+      // The attributes for the lhc-rule-editor component are not 
+      // getting updated as a result. The below steps are used to 
+      // trigger changes to those attributes. 
+      const tmpUserExpressionChoices = this.userExpressionChoices;
+      const tmpCustomExpressionUri = this.customExpressionUri;
+   
+      this.userExpressionChoices = null;
+      this.customExpressionUri = null;
+  
+      this.changeDetectorRef.detectChanges();
+  
+      this.userExpressionChoices = tmpUserExpressionChoices;
+      this.customExpressionUri = tmpCustomExpressionUri;
+    }
   }
 
   /**
    * Check if the Rule Editor can be opened to edit the expression.
-   * @return true if one of the root level checkbox is checked or there is 
-   * a question selected.
+   * @return true if the questionnaire is selected and either the
+   * 'Root level' checkbox or a question is selected.
    */
   canOpenRuleEditor(): boolean {
-    return (this.rootLevel || this.linkId !== null); 
+    return this.fhirQuestionnaire && (this.rootLevel || this.linkId !== null);
   }
 }